requirements.txt
# Django Core
Django==5.2.6
djangorestframework==3.16.1

# Database
psycopg2-binary==2.9.9
pymongo==4.6.1

# Authentication moderne - Alternative à simplejwt
django-rest-knox==4.2.0
social-auth-app-django==5.4.0

# Documentation
drf-spectacular==0.27.0

# Environment variables
django-environ==0.11.2

# Image handling
Pillow==10.1.0

# ASGI/WebSocket support
uvicorn[standard]==0.24.0
channels==4.0.0
daphne==4.0.0

# Development tools
python-decouple==3.8